requirements.txt
# AI Employee - Silver Tier Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies (Bronze Tier)
watchdog>=4.0.0

# Gmail Watcher dependencies
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# LinkedIn Watcher dependencies
playwright>=1.40.0

# Shared utilities
python-dotenv>=1.0.0

# Optional: For better async support
aiohttp>=3.9.0
